+++ src/main/java/cart/application/domain/Order.java
@@ -7,11 +7,12 @@ public class Order {
     private final Long id;
     private final Member member;
     private final List<OrderInfo> orderInfo;
-    private final int originalPrice;
-    private final int usedPoint;
-    private final int pointToAdd;
+    private final long originalPrice;
+    private final long usedPoint;
+    private final long pointToAdd;
 
-    public Order(Long id, Member member, List<OrderInfo> orderInfo, int originalPrice, int usedPoint, int pointToAdd) {
+    public Order(Long id, Member member, List<OrderInfo> orderInfo,
+                 long originalPrice, long usedPoint, long pointToAdd) {
         this.id = id;
         this.member = member;
         this.orderInfo = orderInfo;
@@ -32,15 +33,15 @@ public class Order {
         return orderInfo;
     }
 
-    public int getOriginalPrice() {
+    public long getOriginalPrice() {
         return originalPrice;
     }
 
-    public int getUsedPoint() {
+    public long getUsedPoint() {
         return usedPoint;
     }
 
-    public int getPointToAdd() {
+    public long getPointToAdd() {
         return pointToAdd;
     }
 }
